+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+* Updated pom plugin (2.0.1) containing fix for provided dependencies declared with BOM
+* Revert to old behaviour: in case of gradle plugin project use "maven" publication because its not possible 
+    to differentiate gralde plugin from usual project. In any case, artifacts will be exactly the same everywhere
+    (plugin-publish javadoc and sources tasks will be disabled).  
+
 ### 2.0.0 (2020-01-17)
 * (breaking) Requires gradle 5.1 and above
     - Remove legacy (lazy, without strict publishing) publication configuration